fix: store button type in the component type for better readability
