--- src/components/common/button/types.ts
@@ -1,8 +1,0 @@
-
-export type TButtonProps = React.ButtonHTMLAttributes<HTMLButtonElement> & {
-  children: React.ReactNode;
-  onClick?: () => void;
-  type: 'submit' | 'reset' | 'button';
-  dataTestId: string;
-  className?: string;
-};
